Aura/ai_agency_wealth/requirements.txt
python-dotenv==1.1.1
httpx==0.28.1
duckduckgo-search==5.3.1
qrcode[pil]==8.2
fpdf2==2.8.7
resend==2.23.0
litellm==1.82.0
# ── Quant trading stack ──
ccxt>=4.3.0
requests>=2.31.0
web3>=6.0.0
